Call init / terminator functions in dynamic loader
init_array / fini_array may be present in a loaded binary and contain an
array of functions that need to be called in order to initialize /
terminate a binary. Ensure that the dynamic loader invokes these
functions.

This also requires exposing and handling atexit since it can currently
be invoked by some functions in the init_array. The implementation is
empty for now until it's decided whether CHRE should support this.

Bug: 151847750
Test: Load nanoapps that support init_array / fini_array
Change-Id: Ie959a4d7e50cebc7b786532d61b1b886d4b4ca41