Support RTLD_NEXT as a parameter to dlsym
RTLD_NEXT is used to lookup exported functions inside another library
available on the system. For CHRE's use case, that's limited to
functions exported by the main executable on the given platform that
it's running.

This enables nanoapps to be able to check if a given function exists
instead of requiring that the dynamic loader resolves it as part of
loading the nanoapp which is useful for backwards compatibility.

Bug: 151847750
Test: Load CHRE
Change-Id: If148bf002780cc16b4a2442f4d33435e0b708510