Overridde napp_header flags for TCM-capable nanoapps
TCM-capable nanoapps are nanoapps that are small enough to be placed
into TCM. The new dynamic loader requires a way to know when a nanoapp
should be placed into TCM.

Using the nanoapp header, the daemon can inform the framework of the
nanoapp's flags and the framework can use that to inform the dynamic
loader of where to map the nanoapp into memory.

This also has the added benefit of keeping the dynamic loader unaware of
the details of determining TCM vs. non-TCM placement if it needs to be
used elsewhere.

Bug: 151847750
Test: Compile CHRE and a test nanoapp
Change-Id: Ib6bb28860d2c14929ec3e5f4637e6c412b990b8c