Remove CHRE-specific utilities from flatbuffers
In order to allow multiple components on a single platform to make use
of the CHRE flatbuffers header, remove most of the CHRE-specific logic.
Pull this logic into a utility header that CHRE code can use to ensure
it allocates memory from CHRE's heap and makes use of CHRE-specific util
functions.

The only piece of CHRE-specific code left in the header is CHRE_ASSERT
which is required since it's a macro and it ensures a single assert is
used when the header is utilized by multiple components.

Bug: 160691918
Test: Load CHRE using all CLs in this topic
Change-Id: I8d0276333606315ea792ac99392c70b3693bcf6b