If specified, add a build ID to nanoapps on AoC
The build ID can be used as a fast way to locate the full binary of a
nanoapp when it has been loaded via code download or when triaging a
ramdump.

Since build IDs tend to increment even if a nanoapp hasn't actually
changed, the ID is placed in its own section to allow LLVM to strip its
contents so it can be diffed to see if any logic in the nanoapp changed.

Bug: 161717615
Test: Build nanoapp and verify symbol is placed in section and that
"strings" prints out the symbol value.

Change-Id: I61d4b36a2856a8152bb7474cae0bd09aac61d0f2