Perform fatal error for assert in helper function
Today, the CHRE_ASSERT macro implementation results in a separate string
being created for each invocation to FATAL_ERROR. Instead, if CHRE calls
a separate method that performs the assert and calls FATAL_ERROR with
the assert's filename and line number, there can be significant memory
savings.

With the above change, the incremental size cost of asserts is reduced
by 11k for all platforms resulting in a 6k total incremental cost.

Test: Compile CHRE && ./run_tests.sh && ./run_sim.sh
Change-Id: Ide08f57f2cdec217f871b1d81f45d4ecd60be134